Fix harmless conversion warning in Win32 tray code

+++ src/core/gui/tray_win32.cc
@@ -112,7 +112,7 @@ static bool PrepareIcons(Span<const uint8_t> png, IconSet *out_set)
     IconSet set;
 
     int width, height, channels;
-    uint8_t *img = stbi_load_from_memory(png.ptr, png.len, &width, &height, &channels, 4);
+    uint8_t *img = stbi_load_from_memory(png.ptr, (int)png.len, &width, &height, &channels, 4);
     if (!img) {
         LogError("Failed to load PNG tray icon");
         return false;
